+++ app/renderer/common.js
@@ -265,7 +265,9 @@ const updateScreenshotHighlight = (hide = false) => {
     const url = getUrl()
     const dims = url?.value.split(" ").find(
         arg => arg?.match(/^\d+,\d+,\d+,\d+$/g))
-    const screenCmd = url?.value.replace(/^:/g, "").trim().match(/^screen[a-z]* .*/)
+    const cmd = url?.value.replace(/^:/g, "").trim() ?? ""
+    const screenCmd = cmd.match(/^screenc(opy )?.*/)
+        || cmd.match(/^screens(hot )?.*/)
     const highlight = document.getElementById("screenshot-highlight")
     if (!highlight) {
         return