+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Table of contents to remember the scroll position on hide, until the page is reloaded, for easy toggling
 - Default theme suggest position to be below the tabs
 - Setting "mousevisualmode" now works with double and triple click selections
+- Default screenshot suggestions, highlight behavior and styling to be more prominent
 
 ## Removed